:arrow_up: :pencil2: :globe_with_meridians:
Update submodules and package version

This commit updates the submodules for HomeApiRust and HomeApiWeb, as well as the version of the package.json file to "1.0.6". These changes were made to ensure that the latest versions of the submodules are being used and to reflect the updated version of the package.